requirements.txt
beautifulsoup4==4.8.1
emoji==0.5.4
google==2.0.2
joblib==0.14.0
JPype1==0.7.0
numpy==1.17.4
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.21.3
scipy==1.3.3
six==1.13.0
soupsieve==1.9.5
